requirements.txt
#------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# --- "#"
git+ssh://git@github.com/asuvorov/ddaemon-core-python.git@dev

APScheduler==3.9.1
Pillow==9.4.0

#------------------------------------------------------------------------------
# --- "A"

# -----------------------------------------------------------------------------
# --- "B"

# -----------------------------------------------------------------------------
# --- "C"

# -----------------------------------------------------------------------------
# --- "D"
djangorestframework==3.14.0
djangorestframework-jsonp==1.0.2

# -----------------------------------------------------------------------------
# --- "django-*"
django-admin-rangefilter==0.9.0
django-admin-sortable2==2.1.4
django-bootstrap3-datetimepicker==2.2.3
django-bower==5.2.0
django-compressor==4.3.1
django-cyborg==0.2.1
django-debug-toolbar==3.8.1
django-easy-pdf==0.1.1
django-easy-timezones==0.8.0
django-filter==22.1
django-formset-js==0.5.0
django-geoip-utils==0.1.1
django-grappelli==3.0.4
django-haystack==3.2.1
django-imagekit==4.1.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
django-passwords==0.3.12
django-rest-swagger==2.2.0
django-rosetta==0.9.8
django-secure==1.0.2
django-seo2==1.0.1
django-simple-captcha==0.5.17
django-sslserver==0.22
django-storages==1.13.2
django-taggit==3.1.0
django-taggit-templatetags2==1.6.1
django-timezone-field==5.0
django-twitter-tag==1.2.1
django-url-tools-py3==0.2.1

#------------------------------------------------------------------------------
# --- "E"
elasticsearch==8.6.1

#------------------------------------------------------------------------------
# --- "F"

#------------------------------------------------------------------------------
# --- "G"
geoip2==4.6.0

#------------------------------------------------------------------------------
# --- "H"

#------------------------------------------------------------------------------
# --- "I"

#------------------------------------------------------------------------------
# --- "J"

#------------------------------------------------------------------------------
# --- "K"

#------------------------------------------------------------------------------
# --- "L"

#------------------------------------------------------------------------------
# --- "M"

#------------------------------------------------------------------------------
# --- "N"

#------------------------------------------------------------------------------
# --- "O"

#------------------------------------------------------------------------------
# --- "P"
pendulum==2.1.2
pep8==1.7.1
pep257==0.7.0
pylint==2.15.10
pylint-django==2.5.3

# -----------------------------------------------------------------------------
# --- "python-*"
python-decouple==3.7

#------------------------------------------------------------------------------
# --- "Q"

#------------------------------------------------------------------------------
# --- "R"

# -----------------------------------------------------------------------------
# --- "S"
social-auth-app-django==5.0.0

#------------------------------------------------------------------------------
# --- "T"

#------------------------------------------------------------------------------
# --- "U"

#------------------------------------------------------------------------------
# --- "V"

#------------------------------------------------------------------------------
# --- "W"
whitenoise==6.3.0

#------------------------------------------------------------------------------
# --- "X"

#------------------------------------------------------------------------------
# --- "Y"

#------------------------------------------------------------------------------
# --- "Z"
